refactor: XXXConstants.ktをすべてcom.punyo.slatemap.application.constant直下へ移動
